Possible fix for an issue
This one: https://github.com/lacuna/bifurcan/issues/18

At the very least, it does not cause existing tests to fail, and the
sequence of operations reported in the issue result in correct answers
with these changes, but not without.

If the root node has only one child after removing the last tree node,
then return a new root node that is the one that is its first child.
If any _other_ node that is not the root is in this scenario, do not
change the root node of the tree.