Add consumeseq benchmark, similar to iteration but uses Clojure seq
Also add write-out-csvs action that can be run with a lein command,
which only reads the benchmarks.edn file and writes the CSV files.
Useful if there was anything missing or wrong in that step, and doing
only that again after fixing the problem in the code is much faster
than running benchmarks again.